Learn peer reflexive candidates from all STUN reqs
This patch allows ice4j to learn peer reflexive candidates from all STUN
requests, even if it will respond with a 487 Role Conflict. This allows
peer reflexive candidates to be learned sooner, instead of waiting for
conflict resolution to finish before learning the candidate.

This is necessary for interop with buggy peers that do not switch roles
when they receives the Role Conflict error. Chrome 46.0.2490.86 (current
version) appears to have this problem, which this patch alleviates.

I've looked at the relevant section of the spec (RFC 5244 section
7.2.1), and it appears that this behavior is not prohibited.