ChannelData: rework padding
getDataLength() was returning data length without
padding, so keep current behavior and update the javadoc.
(spotted with findbugs because length in
getDataLength was a dead store)

Introduce padTo4() to round up a number to the next
multiple of 4.

Deprecate encode() and introduce encode(boolean pad)
so we can let the caller decide if he wants padding
(mandatory when using tcp or tls)

See https://tools.ietf.org/html/rfc5766#section-11.5

Signed-off-by: Etienne CHAMPETIER <champetier.etienne@gmail.com>