change MultiplexingXXXSocketSuppport::receive to 'trade' buffers in order to avoid a packet buffer allocation/copy.
Currently, even though the passed-in packet has a buffer already allocated, we allocate a new buffer to receive a packet and put it on the buffer.  When we're ready to fill the given packet with the received data, we copy into the given buffer.
Instead, we can 'trade' the buffers: the passed-in packet's buffer will be used to perform the receive and then added to the queue.  The packet at the front of the queue (which will be discarded anyway) will have its buffer 'stolen' by the passed-in
packet instance. By doing this we can avoid a large buffer allocation and a copy for every packet.